Sdata(BiomassTracking)	R Documentation

Polygon structure and land positions

Description

Example data set of a polygon structure over the Kerguelen plateau. 0 denotes land and positive integer values give the number of the polygon the respective grid point belongs to.

Usage

data(Sdata)

Format

A matrix of integer values.

Source

Thorsten Lenser, just an example.

[Package BiomassTracking version 0.8 Index]