Udata(BiomassTracking)	R Documentation

EW component of a flow field

Description

Example data set of u-components of the flow field over the Kerguelen plateau. Derived from satellite measurements of sea surface height.

Usage

data(Udata)

Format

Three columns, containing longitude, latitude and current speed in m/s. Unknown values are given as -9999.

Source

Belinda Ronai, Australian Antarctic Division, unpublished data.

[Package BiomassTracking version 0.8 Index]